+++ src/app/utils/workflow-visualiser/execution-data-fetcher.js
@@ -8,12 +8,15 @@
  */
 
 import ExecutionDataFetcher from 'onedata-gui-common/utils/workflow-visualiser/execution-data-fetcher';
-import { get, getProperties } from '@ember/object';
+import { get, getProperties, computed } from '@ember/object';
 import OwnerInjector from 'onedata-gui-common/mixins/owner-injector';
 import { inject as service } from '@ember/service';
 import _ from 'lodash';
 import { all as allFulfilled } from 'rsvp';
 import { normalizeWorkflowStatus } from 'onedata-gui-common/utils/workflow-visualiser/statuses';
+
+const notFoundError = { id: 'notFound' };
+const emptyStoreContent = { array: [], isLast: true };
 
 export default ExecutionDataFetcher.extend(OwnerInjector, {
   workflowManager: service(),
@@ -23,6 +26,19 @@ export default ExecutionDataFetcher.extend(OwnerInjector, {
    * @type {Models.AtmWorkflowExecution}
    */
   atmWorkflowExecution: undefined,
+
+  /**
+   * @type {ComputedProperty<Object>}
+   */
+  taskRegistry: computed(
+    'atmWorkflowExecution.lanes',
+    function taskRegistry() {
+      const lanes = this.get('atmWorkflowExecution.lanes');
+      const parallelBoxes = _.flatten(lanes.mapBy('parallelBoxes'));
+      return parallelBoxes.mapBy('taskRegistry')
+        .reduce((gtr, tr) => Object.assign(gtr, tr), {});
+    }
+  ),
 
   /**
    * @override
@@ -36,9 +52,7 @@ export default ExecutionDataFetcher.extend(OwnerInjector, {
 
     const lanes = get(atmWorkflowExecution, 'lanes');
     const parallelBoxes = _.flatten(lanes.mapBy('parallelBoxes'));
-    const taskIds = _.flatten(parallelBoxes.mapBy('taskRegistry')
-      .map(taskRegistry => Object.values(taskRegistry))
-    );
+    const taskIds = Object.values(this.get('taskRegistry'));
     const taskExecutionRecords = await allFulfilled(taskIds.map(taskId =>
       workflowManager.getAtmTaskExecutionById(taskId, { reload: true })
     ));
@@ -72,19 +86,83 @@ export default ExecutionDataFetcher.extend(OwnerInjector, {
    * @override
    */
   async fetchStoreContent(storeSchemaId, startFromIndex, limit, offset) {
-    const {
-      atmWorkflowExecution,
-      workflowManager,
-    } = this.getProperties('atmWorkflowExecution', 'workflowManager');
-
-    const storeRegistry = get(atmWorkflowExecution, 'storeRegistry');
+    const storeRegistry = this.get('atmWorkflowExecution.storeRegistry');
     const storeInstanceId = storeRegistry && storeRegistry[storeSchemaId];
 
-    if (!storeSchemaId) {
-      throw { id: 'notFound' };
+    if (!storeInstanceId) {
+      console.error(
+        'util:workflow-visualiser/execution-data-fetcher#fetchStoreContent: invalid storeSchemaId',
+      );
+      throw notFoundError;
     }
 
-    return workflowManager
+    return await this.fetchStoreInstanceContent(
+      storeInstanceId,
+      startFromIndex,
+      limit,
+      offset
+    );
+  },
+
+  /**
+   * @override
+   */
+  async fetchWorkflowAuditLogContent(startFromIndex, limit, offset) {
+    const auditLogStoreId = this.get('atmWorkflowExecution.systemAuditLogId');
+    // Workflow executions from some alpha releases do not have auditLog stores attached.
+    if (!auditLogStoreId) {
+      return emptyStoreContent;
+    }
+
+    return await this.fetchStoreInstanceContent(
+      auditLogStoreId,
+      startFromIndex,
+      limit,
+      offset
+    );
+  },
+
+  /**
+   * @override
+   */
+  async fetchTaskAuditLogContent(taskSchemaId, startFromIndex, limit, offset) {
+    const {
+      taskRegistry,
+      workflowManager,
+    } = this.getProperties('taskRegistry', 'workflowManager');
+    const taskExecutionId = taskRegistry[taskSchemaId];
+    if (!taskExecutionId) {
+      console.error(
+        'util:workflow-visualiser/execution-data-fetcher#fetchTaskAuditLogContent: invalid taskSchemaId',
+      );
+      throw notFoundError;
+    }
+
+    const taskExecutionRecord =
+      await workflowManager.getAtmTaskExecutionById(taskExecutionId);
+    const auditLogStoreId = get(taskExecutionRecord, 'systemAuditLogId');
+    // Task executions from some alpha releases do not have auditLog stores attached.
+    if (!auditLogStoreId) {
+      return emptyStoreContent;
+    }
+
+    return await this.fetchStoreInstanceContent(
+      auditLogStoreId,
+      startFromIndex,
+      limit,
+      offset
+    );
+  },
+
+  /**
+   * @param {String} storeInstanceId
+   * @param {String} startFromIndex
+   * @param {number} limit
+   * @param {number} offset
+   * @returns {Promise<{array: Array<StoreContentEntry>, isLast: Boolean}>}
+   */
+  async fetchStoreInstanceContent(storeInstanceId, startFromIndex, limit, offset) {
+    return await this.get('workflowManager')
       .getStoreContent(storeInstanceId, startFromIndex, limit, offset);
   },
 });